Moved DriveProfile to lib612
